% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anchormodeling.R
\name{actor.am}
\alias{actor.am}
\title{Example Actor anchor model}
\usage{
actor.am(iteration = 1)
}
\arguments{
\item{iteration}{numeric 1+}
}
\value{
AM class object populated with anchors, attributes, ties and knots from the anchor modeling examples
}
\description{
Example Actor anchor model
}
\details{
Tree iterations of model available: Stage Performances, New Example Model, New startup example in 0.98
}

