% Generated by roxygen2: do not edit by hand
\docType{class}
\name{AM}
\alias{AM}
\title{Anchor Modeling metadata manager}
\format{An R6 class object.}
\usage{
AM
}
\description{
Generate an instance of Anchor Model data warehouse.
}
\examples{
am <- AM$new()
am$log
am$add$A(mne = "AC", desc = "Actor")
am$add$a(anchor = "AC", mne = "GEN", desc = "Gender", knot = "GEN")
am$data
am$log
}
\keyword{datasets}

