% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{AMfun}
\alias{AMfun}
\title{AM object method router}
\usage{
AMfun(class, method)
}
\arguments{
\item{class}{character, one of: \emph{anchor, attribute, tie, knot}.}

\item{method}{character, method name}
}
\value{
function
}
\description{
AM object method router
}
\details{
Used for \code{AMfun(class, "new")(mne, desc, ...)}.
}

