% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/amobj.R
\docType{class}
\name{AMobj}
\alias{AMobj}
\title{AM object}
\format{An R6 class object.}
\usage{
AMobj
}
\description{
Superclass for common features for anchors, attributes, ties and knots.
}
\keyword{datasets}

