% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/amobj.R
\docType{data}
\name{anchor}
\alias{anchor}
\title{Anchor class}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x34f1970> 
 - attr(*, "name")= chr "anchor_generator"
}}
\usage{
anchor
}
\description{
Anchor class
}
\keyword{datasets}

