% Generated by roxygen2: do not edit by hand
\docType{package}
\name{anchormodeling}
\alias{anchormodeling}
\alias{anchormodeling-package}
\title{anchormodeling-package}
\description{
Anchor Modeling metadata manager. AM in-memory Data Warehouse instances. Identity Management instaces. Automated ETL for AM loading.
}
\author{
Jan Gorecki
}
\seealso{
\link{AM}, \link{IM}
}

