% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/amobj.R
\docType{data}
\name{attribute}
\alias{attribute}
\title{Attribute class}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x34d3528> 
 - attr(*, "name")= chr "attribute_generator"
}}
\usage{
attribute
}
\description{
Attribute class
}
\keyword{datasets}

