% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{class.mne}
\alias{class.mne}
\title{Guess class based on mnemonic}
\usage{
class.mne(mne)
}
\arguments{
\item{mne}{character mnemonic in AM naming convention}
}
\value{
character anchor/attribute/tie/knot
}
\description{
Guess class based on mnemonic
}
\details{
Limited only to mnemonics AM naming convention and ties with no more than 2 anchors.
}

