% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{cols.mne}
\alias{cols.mne}
\title{Columns names of 6NF table}
\usage{
cols.mne(mne, desc, class)
}
\arguments{
\item{mne}{character}

\item{desc}{character, not required for anchor and tie}

\item{class}{character optional}
}
\value{
character vector of target generic column names per class
}
\description{
Columns names of 6NF table
}

