% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{get.anchor}
\alias{get.anchor}
\title{Extracts anchor name from mnemonic}
\usage{
get.anchor(mne, i = 1L)
}
\arguments{
\item{mne}{character}

\item{i}{integer location of anchor to extract, for attributes always 1, for ties 1+.}
}
\value{
character mne of anchor
}
\description{
Extracts anchor name from mnemonic
}

