% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/im.R
\docType{class}
\name{IM}
\alias{IM}
\title{Identity Management instance}
\format{An R6 class object.}
\usage{
IM
}
\description{
Keep track on natural keys and generate surrogate keys. It is not part of Anchor Model but complementary identities provider.
}
\examples{
# new IM instance
im <- IM$new()
# define mnemonics and natural keys
im$create(mne = "GEN", nk = "gender")
im$create(mne = "UL", nk = "user_login")
# populate data
data <- data.table(user_login = c("asd","qwe","zxc","asd","zxc","asd"), gender = c("F","M","M","F","M","F"), balance = rnorm(6,10))
data
# update and lookup identities from IM
data.sk <- im$use(data, mne = c("UL","GEN"))
data.sk
# incremental load
data <- data.table(user_login = c("zxc","qaz","qaz","zaq"), gender = c("M","F","F","M"), balance = rnorm(4,10))
data.sk <- im$use(data, mne = c("UL","GEN"))
data.sk
# preview identities
im$ID
# get IM size
print(im)
}
\keyword{datasets}

