% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/amobj.R
\docType{data}
\name{knot}
\alias{knot}
\title{Knot class}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x318bd40> 
 - attr(*, "name")= chr "knot_generator"
}}
\usage{
knot
}
\description{
Knot class
}
\keyword{datasets}

