% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anchormodeling.R
\name{run_anchor_modeler}
\alias{run_anchor_modeler}
\title{Open browser on official Anchor Modeler}
\usage{
run_anchor_modeler()
}
\value{
invisibly NULL
}
\description{
Open browser on official Anchor Modeler
}

