% Generated by roxygen2: do not edit by hand
\docType{data}
\name{SEQ}
\alias{SEQ}
\title{SQL Sequence class}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x32551b0> 
 - attr(*, "name")= chr "SEQ_generator"
}}
\usage{
SEQ
}
\description{
The sequence may be useful when producing unique integers for meta ID on load process
}
\keyword{datasets}

