% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{size.units}
\alias{size.units}
\title{Object size units}
\usage{
size.units(x = getOption("am.size.units"))
}
\arguments{
\item{x}{character, one of \emph{B, KB, MB, GB}. Default \code{options("am.size.units")}.}
}
\value{
integer value to use in \code{bytes/1024^size.units()}.
}
\description{
Format object size numbers.
}

