% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{sub_}
\alias{sub_}
\title{Extracts usually anchor name from mnemonic}
\usage{
sub_(x, i = 1L)
}
\arguments{
\item{x}{character with at least one underscore, for vector param it should expects constant underscore number in each element.}

\item{i}{integer location of character to extract, for attributes always 1, for ties 1+, -1 extracts last element - e.g. knot from tie mnemonic.}
}
\value{
character
}
\description{
Extracts usually anchor name from mnemonic
}

