% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/amobj.R
\docType{data}
\name{tie}
\alias{tie}
\title{Tie class}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x31bbd28> 
 - attr(*, "name")= chr "tie_generator"
}}
\usage{
tie
}
\description{
Tie class
}
\keyword{datasets}

