% Generated by roxygen2: do not edit by hand
\name{actor.data}
\alias{actor.data}
\title{Example data for actor model}
\usage{
actor.data(src = "Stage", iteration = 1L)
}
\arguments{
\item{src}{character \code{"Stage"} or \code{"Actor"} currently.}

\item{iteration}{numeric 1+}
}
\value{
data.table of populated source data load
}
\description{
Example data for actor model
}
\details{
Two iterations of data loads.
}

