% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/am.R
\docType{class}
\name{AM}
\alias{AM}
\title{Anchor Modeling metadata manager}
\format{An R6 class object.}
\usage{
AM
}
\description{
Generate an instance of Anchor Model data warehouse.
}
\examples{
am <- AM$new()
am$log
am$create(class = "anchor", mne = "AC", desc = "Actor")
am$create(class = "attribute", mne = "AC_GEN", desc = "Gender")
am$data
am$log
}
\keyword{datasets}

