% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anchormodeling.R
\docType{data}
\name{AM}
\alias{AM}
\title{Anchor Modeling metadata manager}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x2a00998> 
 - attr(*, "name")= chr "AM_generator"
}}
\usage{
AM
}
\description{
Generate an instance of Anchor Model data warehouse.
}
\examples{
am <- AM$new()
am$log
am$create(class = "anchor", mne = "AC", desc = "Actor")
am$create(class = "attribute", mne = "AC_GEN", desc = "Gender")
am$data
am$log
}
\keyword{datasets}

