% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/amobj.R
\docType{data}
\name{AMobj}
\alias{AMobj}
\title{AM object}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x3520e38> 
 - attr(*, "name")= chr "AMobj_generator"
}}
\usage{
AMobj
}
\description{
Superclass for common features for anchors, attributes, ties and knots.
}
\keyword{datasets}

